move pages into pages folder. resolves #48
