+++ client/src/pages/index/App.svelte
