+++ client/src/pages/index/main.ts
