+++ client/src/pages/profile/Profile.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
 	import '../../tailwind.css';
 
-    import { makeRoutes } from "../../routes";
-    import Alert from '../Alert.svelte';
-    import { AlertColor } from '../../enums';
     import { getContext } from 'svelte';
+    import { makeRoutes } from "../../routes";
+    import { AlertColor } from '../../enums';
+    import Alert from '../../components/Alert.svelte';
 
     const { Get, Patch } = <IHttp>getContext("http");
 	const { Api, Views } = makeRoutes(window.base_path);
