+++ client/src/pages/share/App.svelte
