+++ client/src/pages/share/main.ts
