+++ client/src/pages/wishlist/App.svelte
